requirements.txt
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
apscheduler>=3.10.4
sqlalchemy>=2.0.29
statsmodels>=0.14.0
scikit-learn>=1.4.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
matplotlib>=3.9.0
streamlit>=1.28.0
plotly>=5.17.0
